requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
jinja2
openai